# perf: replace numpy vector store with FAISS for improved query performance