## Summary

This PR replaces the custom numpy-based vector store implementation with FAISS-CPU to achieve significant performance improvements in similarity search operations.

## Motivation

While the numpy implementation was simple and functional, it didn't scale well with large document collections. FAISS (Facebook AI Similarity Search) is a highly optimized library specifically designed for efficient similarity search, providing better performance while maintaining the same functionality.

## Changes

- **Added faiss-cpu as a core dependency**: Updated pyproject.toml to include faiss-cpu>=1.7.4
- **Replaced VectorStore implementation**: Migrated from numpy dot product to FAISS IndexFlatIP
- **Optimized batch operations**: Updated TinyRAG to use batch add method for better FAISS performance
- **Maintained API compatibility**: All existing interfaces remain unchanged

## Performance Results

Benchmark results show significant improvements in query performance:

| Dataset | Metric | Before (numpy) | After (FAISS) | Improvement |
|---------|--------|----------------|---------------|-------------|
| JQaRA   | Query Time | 0.771 sec/query | 0.612 sec/query | **21% faster** |
| JaCWIR  | Query Time | 0.925 sec/query | 0.424 sec/query | **54% faster** |

Accuracy metrics (NDCG, MRR, MAP, Hits) remain unchanged, confirming that the migration preserves search quality.

## Technical Details

- FAISS IndexFlatIP is used for exact cosine similarity search
- Vectors are normalized before indexing to ensure cosine similarity computation
- The implementation maintains the same similarity scoring behavior as the numpy version

## Testing

- [x] All existing tests pass without modification
- [x] Benchmark results confirm performance improvements
- [x] API compatibility is maintained

## Breaking Changes

None. This is a drop-in replacement that maintains full backward compatibility.

🤖 Generated with [Claude Code](https://claude.ai/code)